Remove DenseModel tests
