+++ external/fv3fit/tests/keras/test_sequences.py
@@ -1,0 +1,11 @@
+from fv3fit.keras._models.shared import ThreadedSequencePreLoader
+
+
+def test__ThreadedSequencePreLoader():
+    """ Check correctness of the pre-loaded sequence"""
+    sequence = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+    loader = ThreadedSequencePreLoader(sequence, num_workers=4)
+    result = [item for item in loader]
+    assert len(result) == len(sequence)
+    for item in result:
+        assert item in sequence
